moved spherical throats enpoints, length, piecewise to the _stickAndBall file. [ci skip]
